Add  plugin cpp source so we can load translations
This is a combination of 6 commits.

re-add build-deps

Make .pro file build correctly

This is a combination of 8 commits.

Follow QML plugin example pro

probably need core for QTranslator

The errors are not in the compile, but in the install

we learned something about libdir paths

Fun with paths

wrong dependency

remove all mention of target

Fix installing the lib

Will it blend??